# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Instagram Scalable Image Upload & Metadata Service",
    "description": "\n## Cloud-Native Media Pipeline API\n\nSupports direct-to-S3 presigned multipart uploads, single-table DynamoDB index lookups,\nand non-blocking asynchronous hard purging using AWS SQS.\n    ",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/images/upload-url": {
      "post": {
        "tags": [
          "Image Management"
        ],
        "summary": "Task 1.1: Initiate S3 Presigned Multipart Upload",
        "operationId": "initiate_image_upload_api_v1_images_upload_url_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Returns presigned multipart URLs. Metadata is written after asynchronous validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/images/upload-complete": {
      "post": {
        "tags": [
          "Image Management"
        ],
        "summary": "Complete multipart upload and trigger quarantine validation",
        "operationId": "complete_image_upload_api_v1_images_upload_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/images": {
      "get": {
        "tags": [
          "Image Management"
        ],
        "summary": "Task 1.2: List active images with search filters",
        "operationId": "list_images_api_v1_images_get",
        "parameters": [
          {
            "name": "owner_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter images by Owner ID",
              "title": "Owner Id"
            },
            "description": "Filter images by Owner ID"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter images by Category",
              "title": "Category"
            },
            "description": "Filter images by Category"
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated tags; all tags must match",
              "title": "Tags"
            },
            "description": "Comma-separated tags; all tags must match"
          },
          {
            "name": "filename_prefix",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filename prefix, used with tag",
              "title": "Filename Prefix"
            },
            "description": "Filename prefix, used with tag"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImageItemResponse"
                  },
                  "title": "Response List Images Api V1 Images Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/images/{owner_id}/{image_id}/download": {
      "get": {
        "tags": [
          "Image Management"
        ],
        "summary": "Task 1.3: Get secure view/download URL",
        "operationId": "get_image_download_url_api_v1_images__owner_id___image_id__download_get",
        "parameters": [
          {
            "name": "owner_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Owner Id"
            }
          },
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Image Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/images/{owner_id}/{image_id}": {
      "delete": {
        "tags": [
          "Image Management"
        ],
        "summary": "Task 1.4: Asynchronously delete an image",
        "operationId": "delete_image_api_v1_images__owner_id___image_id__delete",
        "parameters": [
          {
            "name": "owner_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Owner Id"
            }
          },
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Image Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CompletePart": {
        "properties": {
          "PartNumber": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Partnumber"
          },
          "ETag": {
            "type": "string",
            "title": "Etag"
          }
        },
        "type": "object",
        "required": [
          "PartNumber",
          "ETag"
        ],
        "title": "CompletePart"
      },
      "CompleteUploadRequest": {
        "properties": {
          "s3_key": {
            "type": "string",
            "title": "S3 Key"
          },
          "upload_id": {
            "type": "string",
            "title": "Upload Id"
          },
          "parts": {
            "items": {
              "$ref": "#/components/schemas/CompletePart"
            },
            "type": "array",
            "title": "Parts"
          }
        },
        "type": "object",
        "required": [
          "s3_key",
          "upload_id",
          "parts"
        ],
        "title": "CompleteUploadRequest"
      },
      "DownloadResponse": {
        "properties": {
          "image_id": {
            "type": "string",
            "title": "Image Id"
          },
          "download_url": {
            "type": "string",
            "title": "Download Url"
          }
        },
        "type": "object",
        "required": [
          "image_id",
          "download_url"
        ],
        "title": "DownloadResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImageItemResponse": {
        "properties": {
          "image_id": {
            "type": "string",
            "title": "Image Id"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "s3_key": {
            "type": "string",
            "title": "S3 Key"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "image_id",
          "owner_id",
          "category",
          "status",
          "s3_key",
          "created_at"
        ],
        "title": "ImageItemResponse"
      },
      "InitiateUploadRequest": {
        "properties": {
          "owner_id": {
            "type": "string",
            "title": "Owner Id",
            "description": "Unique identifier of the image owner",
            "example": "user_mario_42"
          },
          "file_name": {
            "type": "string",
            "title": "File Name",
            "description": "Original filename of the asset",
            "example": "vacation.jpg"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type",
            "description": "MIME content type",
            "default": "image/jpeg",
            "example": "image/jpeg"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "description": "Category of the image",
            "default": "general",
            "example": "travel"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "description": "Tags used for indexed search",
            "example": [
              "beach",
              "summer"
            ]
          },
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption",
            "description": "Optional text description",
            "example": "Summer in Rome"
          },
          "total_parts": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Total Parts",
            "description": "Number of multipart upload parts requested",
            "default": 1,
            "example": 1
          }
        },
        "type": "object",
        "required": [
          "owner_id",
          "file_name"
        ],
        "title": "InitiateUploadRequest"
      },
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "image_id": {
            "type": "string",
            "title": "Image Id"
          }
        },
        "type": "object",
        "required": [
          "message",
          "image_id"
        ],
        "title": "MessageResponse"
      },
      "PartUrlItem": {
        "properties": {
          "part_number": {
            "type": "integer",
            "title": "Part Number"
          },
          "upload_url": {
            "type": "string",
            "title": "Upload Url"
          }
        },
        "type": "object",
        "required": [
          "part_number",
          "upload_url"
        ],
        "title": "PartUrlItem"
      },
      "UploadResponse": {
        "properties": {
          "image_id": {
            "type": "string",
            "title": "Image Id"
          },
          "upload_id": {
            "type": "string",
            "title": "Upload Id"
          },
          "s3_key": {
            "type": "string",
            "title": "S3 Key"
          },
          "part_urls": {
            "items": {
              "$ref": "#/components/schemas/PartUrlItem"
            },
            "type": "array",
            "title": "Part Urls"
          }
        },
        "type": "object",
        "required": [
          "image_id",
          "upload_id",
          "s3_key",
          "part_urls"
        ],
        "title": "UploadResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "tags": [
    {
      "name": "Image Management",
      "description": "Quarantined multipart uploads, indexed metadata search, secure downloads, and asynchronous deletion."
    }
  ]
}